backend/requirements/development.txt
-r base.txt
pytest==8.3.5
pytest-django==4.9.0
factory-boy==3.3.3